refactor(data): relocate kind schema derives
