+++ crates/aether-actor-derive/Cargo.toml
@@ -16,7 +16,6 @@ syn = { workspace = true }
 aether-data = { path = "../aether-data", features = ["derive"] }
 bytemuck = { workspace = true, features = ["derive"] }
 serde = { workspace = true, features = ["derive", "std"] }
-inventory = { workspace = true }
 aether-actor = { path = "../aether-actor" }
 trybuild = "1"
