+++ crates/aether-actor-derive/src/lib.rs
@@ -1,44 +1,13 @@
-// Derive codegen builds deeply-nested `quote!` trees from `if let Some(...)` branches;
-// `map_or_else` would obscure the control flow. Allow at the crate root because cargo
-// doesn't permit `[lints.clippy]` overrides alongside `lints.workspace = true` in the
-// manifest (iamacoffeepot/aether#854 Phase 1.a).
+// Actor macro codegen builds deeply-nested `quote!` trees from `if let Some(...)`
+// branches; `map_or_else` would obscure the control flow. Allow at the crate
+// root because cargo doesn't permit `[lints.clippy]` overrides alongside
+// `lints.workspace = true` in the manifest (iamacoffeepot/aether#854 Phase 1.a).
 #![allow(clippy::option_if_let_else)]
 
-//! Proc-macro home for `#[derive(Kind)]` and `#[derive(Schema)]` per
-//! ADR-0019 / ADR-0031 / ADR-0032. Kept separate from `aether-data`
-//! because Rust requires proc-macro crates to opt into
-//! `proc-macro = true` and forbids them from exporting non-macro
-//! items; pairing them in the same crate would force every consumer
-//! through the proc-macro toolchain even when they just want the
-//! runtime traits.
-//!
-//! `Kind` emits the `aether_data::Kind` impl (`const NAME`, `const ID`,
-//! optional `const IS_INPUT`) plus the `#[link_section]` statics for
-//! both `aether.kinds` (canonical schema bytes) and
-//! `aether.kinds.labels` (nominal sidecar). The ID is
-//! `fnv1a_64_prefixed(KIND_DOMAIN, canonical_bytes_of(name, schema))`,
-//! matching the substrate-side derivation byte-for-byte (ADR-0030
-//! Phase 2 / ADR-0032). The `KIND_DOMAIN` prefix disjoins the
-//! `Kind::ID` space from `MailboxId` (issue #186). Consumers must also derive (or hand-roll) `Schema`
-//! on the type — the Kind derive walks `<Self as Schema>::SCHEMA`
-//! for canonical bytes and `<Self as Schema>::LABEL_NODE` for
-//! the labels tree.
-//!
-//! `Schema` emits three consts per impl: `SCHEMA` (the `SchemaType`
-//! tree, const-constructible per ADR-0031), `LABEL` (the
-//! `Option<&'static str>` Rust type path from `module_path!()`), and
-//! `LABEL_NODE` (the parallel-shape labels tree the kind's
-//! sidecar record embeds). It also emits `CastEligible` so `repr_c`
-//! flags propagate — field types used as cast-shaped payloads get
-//! eligibility for free without a second derive.
-//!
-//! Field-type handling delegates to `<FieldT as Schema>::SCHEMA` /
-//! `LABEL_NODE` for all cross-crate resolution. The one exception is
-//! `Vec<u8>` — stable Rust forbids the specialization (`Vec<u8>`
-//! would overlap `Vec<T>` because `u8: Schema`), so the derive
-//! pattern-matches the field type's syntax and emits
-//! `SchemaType::Bytes` / `LabelNode::Anonymous` directly when it
-//! sees `Vec<u8>`. Every other shape goes through trait dispatch.
+//! Proc-macro home for the actor SDK attributes: `#[actor]`,
+//! `#[runtime]`, `#[handler]`, `#[fallback]`, `#[capability]`, and
+//! `#[local]`. The data-layer `Kind` / `Schema` derives live in
+//! `aether-data-derive` and are re-exported by `aether-data`.
 
 use std::fs;
 use std::path::Path;
@@ -50,511 +19,9 @@ use syn::meta;
 use syn::parse::Parser;
 use syn::spanned::Spanned;
 use syn::{
-    Attribute, Data, DataEnum, DataStruct, DeriveInput, Expr, ExprLit, Fields, FnArg,
-    GenericArgument, ImplItem, ItemImpl, ItemStruct, Lit, Meta, PathArguments, ReturnType,
-    Signature, Type, parse_macro_input,
+    Attribute, Expr, ExprLit, Fields, FnArg, GenericArgument, ImplItem, ItemImpl, ItemStruct, Lit,
+    Meta, PathArguments, ReturnType, Signature, Type, parse_macro_input,
 };
-
-#[proc_macro_derive(Kind, attributes(kind))]
-pub fn derive_kind(input: TokenStream) -> TokenStream {
-    let input = parse_macro_input!(input as DeriveInput);
-    match expand_kind(&input) {
-        Ok(ts) => ts.into(),
-        Err(e) => e.to_compile_error().into(),
-    }
-}
-
-#[proc_macro_derive(Schema, attributes(kind))]
-pub fn derive_schema(input: TokenStream) -> TokenStream {
-    let input = parse_macro_input!(input as DeriveInput);
-    match expand_schema(&input) {
-        Ok(ts) => ts.into(),
-        Err(e) => e.to_compile_error().into(),
-    }
-}
-
-// Single expansion entry point: emits `Kind` impl, optional
-// `CastEligible`, manifest consts, and retention statics — the surface
-// is wide enough that extracting helpers would force per-helper
-// generic-context arguments without saving readability.
-#[allow(clippy::too_many_lines)]
-fn expand_kind(input: &DeriveInput) -> syn::Result<TokenStream2> {
-    let name = &input.ident;
-    let KindAttr { name: kind_name } = parse_kind_attr(&input.attrs)?;
-    if let Data::Union(u) = &input.data {
-        return Err(syn::Error::new_spanned(
-            u.union_token,
-            "Kind derive does not support unions",
-        ));
-    }
-
-    // ADR-0033 wire-shape autodetect: `#[repr(C)]` on the type means
-    // the substrate carried it as raw cast bytes (and the user has
-    // `#[derive(Pod, Zeroable)]`); anything else is wire-shaped
-    // (ADR-0118 `aether_data::wire`, and the user has
-    // `#[derive(Serialize, Deserialize)]`). The
-    // dispatcher in `#[actor]` calls `Kind::decode_from_bytes` via
-    // `Mail::decode_kind::<K>()`; emitting the body per-impl here is
-    // what lets that one call site compile against types whose Pod /
-    // Deserialize bounds are disjoint.
-    let has_repr_c = struct_has_repr_c(&input.attrs);
-    let decode_body = if has_repr_c {
-        quote! { ::aether_data::__derive_runtime::decode_cast::<Self>(bytes) }
-    } else {
-        quote! { ::aether_data::__derive_runtime::decode_wire::<Self>(bytes) }
-    };
-    // Issue #240: encode mirror. Same `#[repr(C)]` autodetect as
-    // `decode_body` — a single `Sink::send` call site routes through
-    // `Kind::encode_into_bytes`, picking cast or wire at the
-    // kind's derive instead of at every send site.
-    let encode_body = if has_repr_c {
-        quote! { ::aether_data::__derive_runtime::encode_cast::<Self>(self) }
-    } else {
-        quote! { ::aether_data::__derive_runtime::encode_wire::<Self>(self) }
-    };
-
-    // ADR-0032 section emission goes through trait dispatch, not a
-    // syntactic walker. `<Self as Schema>::SCHEMA` / `::LABEL_NODE`
-    // resolve at const-eval after every consumer-side impl is in
-    // scope; the canonical serializers below fold to byte arrays at
-    // compile time. No quiet skips — a type with no `Schema` impl
-    // fails to compile here, which is the behavior ADR-0032 locks
-    // in to keep producer/consumer hashes in lockstep.
-    let upper = to_screaming_snake_case(&name.to_string());
-    let schema_static_ident = quote::format_ident!("__AETHER_SCHEMA_{}", upper);
-    let canonical_len_ident = quote::format_ident!("__AETHER_CANONICAL_LEN_{}", upper);
-    let canonical_bytes_ident = quote::format_ident!("__AETHER_CANONICAL_BYTES_{}", upper);
-    let labels_ident = quote::format_ident!("__AETHER_KIND_LABELS_{}", upper);
-    let labels_len_ident = quote::format_ident!("__AETHER_LABELS_LEN_{}", upper);
-    let labels_bytes_ident = quote::format_ident!("__AETHER_LABELS_BYTES_{}", upper);
-    let kind_static_ident = quote::format_ident!("__AETHER_KIND_MANIFEST_{}", upper);
-    let kind_labels_static_ident = quote::format_ident!("__AETHER_KIND_LABELS_MANIFEST_{}", upper);
-
-    // `#[link_section]` is unsafe under edition 2024 — inert data so
-    // the practical risk is nil, but the `unsafe(...)` wrapper is
-    // required for the attribute to parse. Wasm-target gating keeps
-    // the bytes out of native test executables where they'd just
-    // bloat the binary with no reader.
-    Ok(quote! {
-        impl ::aether_data::Kind for #name {
-            const NAME: &'static str = #kind_name;
-            // ADR-0064: tag the high 4 bits with `Tag::Kind` so kind
-            // ids are distinguishable from mailbox / handle ids by
-            // bit pattern alone. The `KIND_DOMAIN` byte prefix still
-            // rides the FNV input (ADR-0030) — type info ends up
-            // encoded in two independent places that cross-check.
-            // Issue 466: `Kind::ID` is typed `KindId`; the wrapper
-            // wraps the raw `u64` hash. Wire-format sites that need
-            // raw bytes call `.0`; dispatch sites compare `KindId` to
-            // `KindId` directly.
-            const ID: ::aether_data::KindId = ::aether_data::KindId(
-                ::aether_data::with_tag(
-                    ::aether_data::Tag::Kind,
-                    ::aether_data::fnv1a_64_prefixed(
-                        ::aether_data::KIND_DOMAIN,
-                        &#canonical_bytes_ident,
-                    ),
-                ),
-            );
-
-            fn decode_from_bytes(bytes: &[u8]) -> ::core::option::Option<Self> {
-                #decode_body
-            }
-
-            fn encode_into_bytes(&self) -> ::aether_data::__derive_runtime::Vec<u8> {
-                #encode_body
-            }
-        }
-
-        // Intermediate `static` holds the schema value — reading
-        // `<T as Schema>::SCHEMA` by value in a const expression
-        // materializes a temporary whose non-trivial Drop can't run
-        // at compile time. Taking `&SCHEMA_STATIC` sidesteps that
-        // (statics live for the whole program; destructor never runs).
-        static #schema_static_ident: ::aether_data::__derive_runtime::SchemaType =
-            <#name as ::aether_data::Schema>::SCHEMA;
-        const #canonical_len_ident: usize =
-            ::aether_data::__derive_runtime::canonical::canonical_len_kind(
-                #kind_name,
-                &#schema_static_ident,
-            );
-        const #canonical_bytes_ident: [u8; #canonical_len_ident] =
-            ::aether_data::__derive_runtime::canonical::canonical_serialize_kind::<#canonical_len_ident>(
-                #kind_name,
-                &#schema_static_ident,
-            );
-
-        // `static`, not `const`, because `KindLabels` holds `Cow`s
-        // whose non-trivial Drop impl is barred from const-eval.
-        // Statics have program-wide lifetime so the destructor never
-        // needs to run at compile time; const-fn serializers reading
-        // `&#labels_ident` see a stable `'static` reference.
-        static #labels_ident: ::aether_data::__derive_runtime::KindLabels =
-            ::aether_data::__derive_runtime::KindLabels {
-                // Issue 469: `KindLabels.kind_id` is now typed
-                // `KindId` (matches `Kind::ID`); pass through directly.
-                kind_id: <#name as ::aether_data::Kind>::ID,
-                kind_label: ::aether_data::__derive_runtime::Cow::Borrowed(
-                    ::core::concat!(::core::module_path!(), "::", ::core::stringify!(#name)),
-                ),
-                root: <#name as ::aether_data::Schema>::LABEL_NODE,
-            };
-        const #labels_len_ident: usize =
-            ::aether_data::__derive_runtime::canonical::canonical_len_labels(&#labels_ident);
-        const #labels_bytes_ident: [u8; #labels_len_ident] =
-            ::aether_data::__derive_runtime::canonical::canonical_serialize_labels::<#labels_len_ident>(
-                &#labels_ident,
-            );
-
-        // ADR-0028 / ADR-0032 / ADR-0118: `aether.kinds` ships
-        // `[version_byte][canonical_bytes]`, where the canonical bytes are
-        // now the owned aether-wire encoding (issue 1984) — so every
-        // `Kind::ID` regenerates, gated loudly behind this version byte.
-        // The version byte is `aether_data::KINDS_SECTION_VERSION`, the
-        // single source of truth the reader also reads.
-        #[cfg(target_family = "wasm")]
-        #[used]
-        #[unsafe(link_section = "aether.kinds")]
-        static #kind_static_ident: [u8; #canonical_len_ident + 1] = {
-            let mut out = [0u8; #canonical_len_ident + 1];
-            out[0] = ::aether_data::KINDS_SECTION_VERSION;
-            let mut i = 0;
-            while i < #canonical_len_ident {
-                out[i + 1] = #canonical_bytes_ident[i];
-                i += 1;
-            }
-            out
-        };
-
-        #[cfg(target_family = "wasm")]
-        #[used]
-        #[unsafe(link_section = "aether.kinds.labels")]
-        static #kind_labels_static_ident: [u8; #labels_len_ident + 1] = {
-            let mut out = [0u8; #labels_len_ident + 1];
-            // ADR-0118 / issue 1984: the labels record is the owned
-            // aether-wire encoding of `KindLabels`. v0x03 made records
-            // self-identifying (`kind_id`); the reader still pairs by id.
-            // The version byte is `aether_data::LABELS_SECTION_VERSION`,
-            // the single source of truth the reader also reads.
-            out[0] = ::aether_data::LABELS_SECTION_VERSION;
-            let mut i = 0;
-            while i < #labels_len_ident {
-                out[i + 1] = #labels_bytes_ident[i];
-                i += 1;
-            }
-            out
-        };
-
-        // Issue #243: native-side auto-collection. The wasm
-        // `aether.kinds` custom-section above carries the canonical
-        // bytes for guest-side discovery; on native, the substrate's
-        // `descriptors::all()` materializes the Hub-shipped
-        // `KindDescriptor` list by iterating these inventory entries.
-        // Cfg-gated to non-wasm targets because `inventory` doesn't
-        // link on `wasm32-unknown-unknown`.
-        #[cfg(not(target_family = "wasm"))]
-        ::aether_data::__inventory::inventory::submit! {
-            ::aether_data::__inventory::DescriptorEntry {
-                name: <#name as ::aether_data::Kind>::NAME,
-                schema: &#schema_static_ident,
-            }
-        }
-    })
-}
-
-fn cast_eligible_expr_for_struct(has_repr_c: bool, fields: &[FieldInfo]) -> TokenStream2 {
-    if !has_repr_c {
-        return quote! { false };
-    }
-    if fields.is_empty() {
-        return quote! { true };
-    }
-    let parts = fields.iter().map(|f| {
-        let ty = &f.ty;
-        quote! { <#ty as ::aether_data::CastEligible>::ELIGIBLE }
-    });
-    quote! { #(#parts)&&* }
-}
-
-fn expand_schema(input: &DeriveInput) -> syn::Result<TokenStream2> {
-    let name = &input.ident;
-    let name_str = name.to_string();
-    let (body, label_node_body, cast_eligible_expr) = match &input.data {
-        Data::Struct(_) => {
-            let fields = struct_fields(input)?;
-            let has_repr_c = struct_has_repr_c(&input.attrs);
-            (
-                expand_schema_struct(&fields)?,
-                expand_label_node_struct(&name_str, &fields),
-                cast_eligible_expr_for_struct(has_repr_c, &fields),
-            )
-        }
-        Data::Enum(e) => (
-            expand_schema_enum(e)?,
-            expand_label_node_enum(&name_str, e),
-            quote! { false },
-        ),
-        Data::Union(u) => {
-            return Err(syn::Error::new_spanned(
-                u.union_token,
-                "Schema derive does not support unions",
-            ));
-        }
-    };
-    Ok(quote! {
-        impl ::aether_data::Schema for #name {
-            const SCHEMA: ::aether_data::__derive_runtime::SchemaType = #body;
-            const LABEL: ::core::option::Option<&'static str> = ::core::option::Option::Some(
-                ::core::concat!(::core::module_path!(), "::", ::core::stringify!(#name)),
-            );
-            const LABEL_NODE: ::aether_data::__derive_runtime::LabelNode = #label_node_body;
-        }
-
-        impl ::aether_data::CastEligible for #name {
-            const ELIGIBLE: bool = #cast_eligible_expr;
-        }
-    })
-}
-
-/// Emit the `LabelNode::Struct` literal for the type's `LABEL_NODE`
-/// const. Field names come from the Rust source; nested-field label
-/// nodes resolve via `<FieldT as Schema>::LABEL_NODE` trait dispatch.
-/// `Vec<u8>` field specialization: the schema side reports `Bytes`,
-/// the labels side reports `Anonymous` (no nominal info for a raw
-/// byte buffer).
-fn expand_label_node_struct(type_ident: &str, fields: &[FieldInfo]) -> TokenStream2 {
-    let field_names = fields.iter().enumerate().map(|(idx, f)| match &f.ident {
-        Some(id) => id.to_string(),
-        None => idx.to_string(),
-    });
-    let field_name_entries = field_names.map(|n| {
-        quote! { ::aether_data::__derive_runtime::Cow::Borrowed(#n) }
-    });
-    let field_node_exprs = fields.iter().map(|f| field_label_node_expr(&f.ty));
-    quote! {
-        ::aether_data::__derive_runtime::LabelNode::Struct {
-            type_label: ::core::option::Option::Some(
-                ::aether_data::__derive_runtime::Cow::Borrowed(
-                    ::core::concat!(::core::module_path!(), "::", #type_ident),
-                ),
-            ),
-            field_names: ::aether_data::__derive_runtime::Cow::Borrowed(&[
-                #( #field_name_entries ),*
-            ]),
-            fields: ::aether_data::__derive_runtime::Cow::Borrowed(&[
-                #( #field_node_exprs ),*
-            ]),
-        }
-    }
-}
-
-fn expand_label_node_enum(type_ident: &str, data: &DataEnum) -> TokenStream2 {
-    let variant_entries = data.variants.iter().map(|v| {
-        let vname = v.ident.to_string();
-        match &v.fields {
-            Fields::Unit => quote! {
-                ::aether_data::__derive_runtime::VariantLabel::Unit {
-                    name: ::aether_data::__derive_runtime::Cow::Borrowed(#vname),
-                }
-            },
-            Fields::Unnamed(unnamed) => {
-                let field_exprs = unnamed.unnamed.iter().map(|f| field_label_node_expr(&f.ty));
-                quote! {
-                    ::aether_data::__derive_runtime::VariantLabel::Tuple {
-                        name: ::aether_data::__derive_runtime::Cow::Borrowed(#vname),
-                        fields: ::aether_data::__derive_runtime::Cow::Borrowed(&[
-                            #( #field_exprs ),*
-                        ]),
-                    }
-                }
-            }
-            Fields::Named(named) => {
-                let field_name_entries = named.named.iter().map(|f| {
-                    let fname = f
-                        .ident
-                        .as_ref()
-                        .map(ToString::to_string)
-                        .unwrap_or_default();
-                    quote! { ::aether_data::__derive_runtime::Cow::Borrowed(#fname) }
-                });
-                let field_node_exprs = named.named.iter().map(|f| field_label_node_expr(&f.ty));
-                quote! {
-                    ::aether_data::__derive_runtime::VariantLabel::Struct {
-                        name: ::aether_data::__derive_runtime::Cow::Borrowed(#vname),
-                        field_names: ::aether_data::__derive_runtime::Cow::Borrowed(&[
-                            #( #field_name_entries ),*
-                        ]),
-                        fields: ::aether_data::__derive_runtime::Cow::Borrowed(&[
-                            #( #field_node_exprs ),*
-                        ]),
-                    }
-                }
-            }
-        }
-    });
-    quote! {
-        ::aether_data::__derive_runtime::LabelNode::Enum {
-            type_label: ::core::option::Option::Some(
-                ::aether_data::__derive_runtime::Cow::Borrowed(
-                    ::core::concat!(::core::module_path!(), "::", #type_ident),
-                ),
-            ),
-            variants: ::aether_data::__derive_runtime::Cow::Borrowed(&[
-                #( #variant_entries ),*
-            ]),
-        }
-    }
-}
-
-/// Expression for a field's `LabelNode` — trait dispatch through
-/// `<T as Schema>::LABEL_NODE` for most types. `Vec<u8>` is the one
-/// exception (pattern-matched just like `field_type_schema_expr`)
-/// and maps to `LabelNode::Anonymous` because `Bytes` carries no
-/// structural children to label.
-fn field_label_node_expr(ty: &Type) -> TokenStream2 {
-    if is_vec_u8(ty) {
-        quote! { ::aether_data::__derive_runtime::LabelNode::Anonymous }
-    } else {
-        quote! { <#ty as ::aether_data::Schema>::LABEL_NODE }
-    }
-}
-
-fn expand_schema_struct(fields: &[FieldInfo]) -> syn::Result<TokenStream2> {
-    if fields.is_empty() {
-        return Ok(quote! { ::aether_data::__derive_runtime::SchemaType::Unit });
-    }
-
-    for f in fields {
-        reject_hashmap(&f.ty)?;
-    }
-
-    let entries = fields.iter().enumerate().map(|(idx, f)| {
-        let name = match &f.ident {
-            Some(id) => id.to_string(),
-            // Tuple struct field — name positionally so the hub still
-            // has something to render in `describe_kinds`. The wire
-            // format doesn't care; field names are advisory metadata.
-            None => idx.to_string(),
-        };
-        let ty_expr = field_type_schema_expr(&f.ty);
-        quote! {
-            ::aether_data::__derive_runtime::NamedField {
-                name: ::aether_data::__derive_runtime::Cow::Borrowed(#name),
-                ty: #ty_expr,
-            }
-        }
-    });
-
-    Ok(quote! {
-        ::aether_data::__derive_runtime::SchemaType::Struct {
-            fields: ::aether_data::__derive_runtime::Cow::Borrowed(&[ #( #entries ),* ]),
-            repr_c: <Self as ::aether_data::CastEligible>::ELIGIBLE,
-        }
-    })
-}
-
-fn expand_schema_enum(data: &DataEnum) -> syn::Result<TokenStream2> {
-    for v in &data.variants {
-        for f in &v.fields {
-            reject_hashmap(&f.ty)?;
-        }
-    }
-
-    let variant_entries = data.variants.iter().enumerate().map(|(idx, v)| {
-        let name = v.ident.to_string();
-        // Enum variants past `u32::MAX` aren't a realistic schema; the
-        // canonical-bytes wire format stores discriminants as u32.
-        #[allow(clippy::cast_possible_truncation)]
-        let discriminant = idx as u32;
-        match &v.fields {
-            Fields::Unit => quote! {
-                ::aether_data::__derive_runtime::EnumVariant::Unit {
-                    name: ::aether_data::__derive_runtime::Cow::Borrowed(#name),
-                    discriminant: #discriminant,
-                }
-            },
-            Fields::Unnamed(unnamed) => {
-                let field_exprs = unnamed
-                    .unnamed
-                    .iter()
-                    .map(|f| field_type_schema_expr(&f.ty));
-                quote! {
-                    ::aether_data::__derive_runtime::EnumVariant::Tuple {
-                        name: ::aether_data::__derive_runtime::Cow::Borrowed(#name),
-                        discriminant: #discriminant,
-                        fields: ::aether_data::__derive_runtime::Cow::Borrowed(&[ #( #field_exprs ),* ]),
-                    }
-                }
-            }
-            Fields::Named(named) => {
-                let field_exprs = named.named.iter().map(|f| {
-                    let fname = f.ident.as_ref().map(ToString::to_string).unwrap_or_default();
-                    let ty_expr = field_type_schema_expr(&f.ty);
-                    quote! {
-                        ::aether_data::__derive_runtime::NamedField {
-                            name: ::aether_data::__derive_runtime::Cow::Borrowed(#fname),
-                            ty: #ty_expr,
-                        }
-                    }
-                });
-                quote! {
-                    ::aether_data::__derive_runtime::EnumVariant::Struct {
-                        name: ::aether_data::__derive_runtime::Cow::Borrowed(#name),
-                        discriminant: #discriminant,
-                        fields: ::aether_data::__derive_runtime::Cow::Borrowed(&[ #( #field_exprs ),* ]),
-                    }
-                }
-            }
-        }
-    });
-
-    Ok(quote! {
-        ::aether_data::__derive_runtime::SchemaType::Enum {
-            variants: ::aether_data::__derive_runtime::Cow::Borrowed(&[ #( #variant_entries ),* ]),
-        }
-    })
-}
-
-// Pattern-match `Vec<u8>` at the field-type level so it lands as
-// `SchemaType::Bytes` rather than the generic `Vec(Scalar(U8))`. Every
-// other shape delegates to the `Schema` trait's const — wrapped in
-// `SchemaCell::Static` at recursive positions so the literal stays
-// const-constructible.
-fn field_type_schema_expr(ty: &Type) -> TokenStream2 {
-    if is_vec_u8(ty) {
-        quote! { ::aether_data::__derive_runtime::SchemaType::Bytes }
-    } else {
-        quote! { <#ty as ::aether_data::Schema>::SCHEMA }
-    }
-}
-
-/// Returns `true` when `ty` is syntactically a `Vec<u8>` (or any qualified
-/// spelling whose outer type ends in `Vec` and whose element type's last
-/// segment is `u8`, e.g. `Vec<core::primitive::u8>`). The check is purely
-/// syntactic — a type alias (`type Blob = Vec<u8>`) is not resolved by the
-/// proc macro and falls through to the generic `Vec` schema.
-fn is_vec_u8(ty: &Type) -> bool {
-    let Type::Path(tp) = ty else { return false };
-    let Some(seg) = tp.path.segments.last() else {
-        return false;
-    };
-    if seg.ident != "Vec" {
-        return false;
-    }
-    let PathArguments::AngleBracketed(args) = &seg.arguments else {
-        return false;
-    };
-    let Some(GenericArgument::Type(Type::Path(inner))) = args.args.first() else {
-        return false;
-    };
-    // Match on the last segment so qualified spellings like
-    // `core::primitive::u8` are recognized alongside bare `u8`.
-    inner
-        .path
-        .segments
-        .last()
-        .is_some_and(|seg| seg.ident == "u8")
-}
 
 /// Walk a field-type syntactic tree and reject `HashMap` anywhere
 /// inside it (issue #232). `HashMap`'s iteration order is hash-state-
@@ -620,94 +87,6 @@ fn reject_hashmap(ty: &Type) -> syn::Result<()> {
         _ => {}
     }
     Ok(())
-}
-
-struct KindAttr {
-    name: String,
-}
-
-fn parse_kind_attr(attrs: &[Attribute]) -> syn::Result<KindAttr> {
-    for attr in attrs {
-        if !attr.path().is_ident("kind") {
-            continue;
-        }
-        let mut name: Option<String> = None;
-        attr.parse_nested_meta(|meta| {
-            if meta.path.is_ident("name") {
-                let value = meta.value()?;
-                let expr: Expr = value.parse()?;
-                if let Expr::Lit(lit) = &expr
-                    && let Lit::Str(s) = &lit.lit
-                {
-                    name = Some(s.value());
-                    return Ok(());
-                }
-                return Err(meta.error("`name` must be a string literal"));
-            }
-            Err(meta.error("expected `name = \"...\"`"))
-        })?;
-        if let Some(name) = name {
-            return Ok(KindAttr { name });
-        }
-    }
-    Err(syn::Error::new(
-        attrs
-            .first()
-            .map_or_else(proc_macro2::Span::call_site, Spanned::span),
-        "missing `#[kind(name = \"...\")]` attribute",
-    ))
-}
-
-fn struct_has_repr_c(attrs: &[Attribute]) -> bool {
-    for attr in attrs {
-        if !attr.path().is_ident("repr") {
-            continue;
-        }
-        let Meta::List(list) = &attr.meta else {
-            continue;
-        };
-        let mut has_c = false;
-        let _ = list.parse_nested_meta(|meta| {
-            if meta.path.is_ident("C") {
-                has_c = true;
-            }
-            Ok(())
-        });
-        if has_c {
-            return true;
-        }
-    }
-    false
-}
-
-fn struct_fields(input: &DeriveInput) -> syn::Result<Vec<FieldInfo>> {
-    let Data::Struct(DataStruct { fields, .. }) = &input.data else {
-        return Err(syn::Error::new_spanned(&input.ident, "expected struct"));
-    };
-    Ok(match fields {
-        Fields::Named(named) => named
-            .named
-            .iter()
-            .map(|f| FieldInfo {
-                ident: f.ident.clone(),
-                ty: f.ty.clone(),
-            })
-            .collect(),
-        Fields::Unnamed(unnamed) => unnamed
-            .unnamed
-            .iter()
-            .map(|f| FieldInfo {
-                ident: None,
-                ty: f.ty.clone(),
-            })
-            .collect(),
-        Fields::Unit => Vec::new(),
-    })
-}
-
-pub(crate) struct FieldInfo {
-    pub(crate) ident: Option<syn::Ident>,
-    pub(crate) ty: Type,
 }
 
 fn to_screaming_snake_case(s: &str) -> String {
@@ -4066,8 +3445,8 @@ fn build_kinds_section_retention_statics(
             idx
         );
         quote! {
-            // Mirrors the intermediate-static pattern in `expand_kind`
-            // (mail-derive/src/lib.rs) so const-eval of the serializer
+            // Mirrors the intermediate-static pattern in the Kind derive
+            // (`aether-data-derive`) so const-eval of the serializer
             // sees a `&'static SchemaType` / `&'static KindLabels`
             // instead of materializing a temporary whose non-trivial
             // Drop can't run at compile time.
@@ -4083,7 +3462,7 @@ fn build_kinds_section_retention_statics(
                     <#k as ::aether_actor::__macro_internals::Kind>::NAME,
                     &#schema_ident,
                 );
-            // Same wire shape as `expand_kind`'s primary emission
+            // Same wire shape as the Kind derive's primary emission
             // (ADR-0118 / issue 1984: the owned aether-wire encoding) so
             // retention records (when this kind lives in a dependency
             // rlib) pair cleanly with the primary records by id.
@@ -4138,7 +3517,7 @@ fn build_kinds_section_retention_statics(
             static #labels_section_ident: [u8; #labels_len_ident + 1] = {
                 let mut out = [0u8; #labels_len_ident + 1];
                 // ADR-0118 / issue 1984: the owned aether-wire
-                // encoding of `KindLabels`, matching `expand_kind`.
+                // encoding of `KindLabels`, matching the Kind derive.
                 out[0] = ::aether_actor::__macro_internals::LABELS_SECTION_VERSION;
                 let mut i = 0;
                 while i < #labels_len_ident {
