+++ crates/aether-data/Cargo.toml
@@ -5,18 +5,12 @@ edition.workspace = true
 license.workspace = true
 
 [features]
-# Re-export `#[derive(Kind)]` and `#[derive(Schema)]` from
-# `aether-actor-derive` (ADR-0077 / issue 552 stage 0: the kind +
-# actor proc-macros consolidated into one crate so the derive home
-# matches the SDK home), plus the `#[transform]` attribute macro from
-# `aether-data-derive` (ADR-0048 §1: a transform is a pure data-layer
-# primitive, so its macro re-exports from the data layer). Off by
-# default so guests that hand-write `impl Kind for ...` don't pay the
-# proc-macro compile cost.
-derive = ["dep:aether-actor-derive", "dep:aether-data-derive"]
+# Re-export `#[derive(Kind)]`, `#[derive(Schema)]`, and `#[transform]`
+# from `aether-data-derive`. Off by default so guests that hand-write
+# `impl Kind for ...` don't pay the proc-macro compile cost.
+derive = ["dep:aether-data-derive"]
 
 [dependencies]
-aether-actor-derive = { path = "../aether-actor-derive", optional = true }
 aether-data-derive = { path = "../aether-data-derive", optional = true }
 bytemuck = { workspace = true, features = ["derive"] }
 serde = { workspace = true, default-features = false, features = ["derive", "alloc"] }
